more code consolidation
