+++ components/ProposalVotingPower/LockedCommunityVotingPower.tsx
@@ -18,7 +18,7 @@ import { useRealmCommunityMintInfoQuery } from '@hooks/queries/mintInfo'
 import { useConnection } from '@solana/wallet-adapter-react'
 import BN from 'bn.js'
 import { useVsrGovpower } from '@hooks/queries/plugins/vsr'
-import VotingPowerBox from 'VoteStakeRegistry/components/TokenBalance/VotingPowerBox'
+import VSRCommunityVotingPower from 'VoteStakeRegistry/components/TokenBalance/VSRVotingPower'
 
 interface Props {
   className?: string
@@ -31,15 +31,11 @@ export default function LockedCommunityVotingPower(props: Props) {
   const { realmInfo, realmTokenAccount } = useRealm()
   const client = useVotePluginsClientStore((s) => s.state.vsrClient)
   const { connection } = useConnection()
-  const deposits = useDepositStore((s) => s.state.deposits)
-
   const endpoint = connection.rpcEndpoint
 
   const getOwnedDeposits = useDepositStore((s) => s.getOwnedDeposits)
   const votingPower = useVsrGovpower().result?.result ?? new BN(0)
-  const votingPowerFromDeposits = useDepositStore(
-    (s) => s.state.votingPowerFromDeposits
-  )
+
   const wallet = useWalletOnePointOh()
   const isLoading = useDepositStore((s) => s.state.isLoading)
 
@@ -49,12 +45,6 @@ export default function LockedCommunityVotingPower(props: Props) {
   const tokenOwnerRecordPk = currentTokenOwnerRecord
     ? currentTokenOwnerRecord.pubkey
     : null
-
-  const depositRecord = deposits.find(
-    (deposit) =>
-      deposit.mint.publicKey.toBase58() ===
-        realm?.account.communityMint.toBase58() && deposit.lockup.kind.none
-  )
 
   const depositMint = realm?.account.communityMint
   const depositAmount = realmTokenAccount
@@ -68,29 +58,6 @@ export default function LockedCommunityVotingPower(props: Props) {
     votingPower && mint
       ? getMintDecimalAmount(mint, votingPower)
       : new BigNumber('0')
-
-  const tokenAmount =
-    depositRecord && mint
-      ? new BigNumber(
-          getMintDecimalAmount(mint, depositRecord.amountDepositedNative)
-        )
-      : new BigNumber('0')
-
-  const lockedTokensAmount = mint
-    ? deposits
-        .filter(
-          (x) =>
-            typeof x.lockup.kind['none'] === 'undefined' &&
-            x.mint.publicKey.toBase58() ===
-              realm?.account.communityMint.toBase58()
-        )
-        .reduce(
-          (curr, next) =>
-            curr.plus(new BigNumber(next.currentlyLocked.toString())),
-          new BigNumber(0)
-        )
-        .shiftedBy(-mint.decimals)
-    : new BigNumber('0')
 
   const deposit = useCallback(async () => {
     if (
@@ -152,30 +119,7 @@ export default function LockedCommunityVotingPower(props: Props) {
           You do not have any voting power in this dao.
         </div>
       ) : (
-        <>
-          <VotingPowerBox
-            votingPower={votingPower}
-            mint={mint}
-            votingPowerFromDeposits={votingPowerFromDeposits}
-            className="p-3"
-          />
-          <div className="pt-4 px-4">
-            <p className="flex mb-1.5 text-xs">
-              <span>{tokenName} Deposited</span>
-              <span className="font-bold ml-auto text-fgd-1">
-                {tokenAmount.isNaN() ? '0' : tokenAmount.toFormat()}
-              </span>
-            </p>
-            <p className="flex text-xs">
-              <span>{tokenName} Locked</span>
-              <span className="font-bold ml-auto text-fgd-1">
-                {lockedTokensAmount.isNaN()
-                  ? '0'
-                  : lockedTokensAmount.toFormat()}
-              </span>
-            </p>
-          </div>
-        </>
+        <VSRCommunityVotingPower />
       )}
       {depositAmount.isGreaterThan(0) && (
         <>
